fix: no logged in but go dashboard
